dev-requirements.txt
pytest==5.0.1
autopep8==1.4.4
rope==0.14.0
pylint==2.3.1
pylint-django==2.0.10
pytest-django==3.6.0
